requirements.txt
# 文献分析与综述助手 - 后端依赖（请在虚拟环境中安装）
# 创建虚拟环境: python -m venv .venv
# 激活: Windows .venv\Scripts\activate   Linux/Mac source .venv/bin/activate
# 安装: pip install -r requirements.txt -i https://pypi.org/simple --trusted-host pypi.org --trusted-host files.pythonhosted.org

# Web
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0

# PDF
pdfminer.six>=20221105

# Elasticsearch（BM25 检索）
elasticsearch>=8.0.0

# 向量检索（sentence-transformers + FAISS）
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# 分析与 NLP
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
keybert>=0.8.0
transformers>=4.30.0
torch>=2.0.0

# 智谱 GLM-4（OpenAI 兼容）
openai>=1.0.0
